+++ packages/typespec-rust/src/tcgcadapter/adapter.ts
@@ -167,6 +167,8 @@ export class Adapter {
       this.crate.enums.push(rustEnum);
     }
 
+    const processedTypes = new Set<string>();
+
     for (const model of this.ctx.sdkPackage.models) {
       if ((model.usage & tcgc.UsageFlags.Input) === 0 && (model.usage & tcgc.UsageFlags.Output) === 0 && (model.usage & tcgc.UsageFlags.Spread) === 0) {
         // skip types without input and output usage. this will include core
@@ -174,6 +176,15 @@ export class Adapter {
         // we keep the models for spread params as we internally use them.
         continue;
       }
+
+      // TODO: workaround for https://github.com/Azure/typespec-azure/issues/3614
+      if (processedTypes.has(model.name)) {
+        continue;
+      } else {
+        processedTypes.add(model.name);
+      }
+      // END workaround
+
       const rustModel = this.getModel(model);
       this.crate.models.push(rustModel);
     }
@@ -1650,13 +1661,14 @@ export class Adapter {
     // adapt the response headers and add them to the trait
     for (const header of responseHeaders) {
       let responseHeader: rust.ResponseHeader;
+      const lowerCasedHeader = header.serializedName.toLowerCase();
       if (header.type.kind === 'dict') {
         if (header.serializedName !== 'x-ms-meta' && header.serializedName !== 'x-ms-or') {
           throw new AdapterError('InternalError', `unexpected response header collection ${header.serializedName}`, header.__raw.node);
         }
-        responseHeader = new rust.ResponseHeaderHashMap(snakeCaseName(header.name), header.serializedName);
+        responseHeader = new rust.ResponseHeaderHashMap(snakeCaseName(header.name), lowerCasedHeader);
       } else {
-        responseHeader = new rust.ResponseHeaderScalar(snakeCaseName(header.name), fixETagName(header.serializedName), this.typeToWireType(this.getType(header.type)));
+        responseHeader = new rust.ResponseHeaderScalar(snakeCaseName(header.name), fixETagName(lowerCasedHeader), this.typeToWireType(this.getType(header.type)));
       }
 
       responseHeader.docs = this.adaptDocs(header.summary, header.doc);