+++ packages/typespec-rust/CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fixed various cases of bad codegen.
   * Referencing optional client parameters.
   * Query parameters that are exploded arrays.
+  * Deduplicate parameter name that collides with the `options` parameter.
 
 ### Features Added